PORT-018: Isolating MapForm and Loading components
