+++ web/src/components/GeoFootball/index.tsx
@@ -1,13 +1,14 @@
 import type { Station, GroupedStations } from './types'
-import { MapContainer, TileLayer, Marker, Popup, useMapEvents } from 'react-leaflet'
-import { Icon, LeafletEvent, Marker as LeafletMarker } from 'leaflet'
+import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet'
+import { Icon } from 'leaflet'
 import MarkerClusterGroup from 'react-leaflet-cluster'
-import { useState } from 'react'
-import { useForm } from 'react-hook-form'
-import { deleteStation, saveStation } from './slice'
 import Image from 'next/image'
-
+import { useState } from 'react'
 import { Noto_Color_Emoji } from 'next/font/google'
+
+import MapForm from './form'
+import { deleteStation, flag } from './slice'
+
 import styles from './index.module.css'
 
 
@@ -17,152 +18,11 @@ interface Properties {
 	stationsByCountry: GroupedStations,
 }
 
-function flag(country: string) {
-	const points = country
-		.toUpperCase()
-		.split('')
-		.map((char) => 127397 + char.charCodeAt(0));
-	return String.fromCodePoint(...points);
-}
-
 const train = new Icon({
 	iconUrl: 'https://cdn-icons-png.flaticon.com/512/1702/1702305.png',
 	iconSize: [24, 24],
 	iconAnchor: [12, 12],
 })
-
-const pin = new Icon({
-	iconUrl: 'https://cdn-icons-png.flaticon.com/512/2684/2684860.png',
-	iconSize: [24, 24],
-	iconAnchor: [0, 24],
-})
-
-interface MapFormProperties {
-	clusters: GroupedStations,
-	setClusters: Function,
-}
-
-interface LoadingProperties {
-	text: string,
-}
-
-const Loading = ({ text }: LoadingProperties) => {
-	return (
-		<>
-			<Image src="https://i.gifer.com/VAyR.gif" alt={text} width={16} height={16} /> {text}
-		</>
-	)
-}
-
-const MapForm = ({ clusters, setClusters }: MapFormProperties) => {
-	const countries = Object.keys(clusters)
-	const initialStation = {
-		name: '',
-		slug: '',
-		country: 'GB',
-		time_zone: 'Europe/London',
-		latitude: 51.478,
-		longitude: 0,
-	} as Station
-	const [station, setStation] = useState(initialStation)
-  const [marker, setMarker] = useState(null);
-
-	const {
-		register,
-		handleSubmit,
-		formState: { isSubmitting, errors },
-		reset,
-		setError,
-	} = useForm({
-		defaultValues: { ...initialStation },
-	});
-
-	useMapEvents({
-		dblclick(event) {
-			setStation({
-				...station,
-				latitude: event.latlng.lat,
-				longitude: event.latlng.lng,
-			})
-      marker && (marker as LeafletMarker).openPopup()
-		}
-	})
-
-	const save = async (data: object) => {
-		const response = await saveStation({
-			...data,
-			latitude: station.latitude,
-			longitude: station.longitude,
-		} as Station)
-		if (response.ok) {
-      marker && (marker as LeafletMarker).closePopup()
-			const record = await response.json() as Station
-			setStation({ ...initialStation });
-			const cluster = clusters[record.country] as Station[]
-			cluster.push(record)
-
-			setClusters({
-				...clusters,
-				[record.country]: [...cluster],
-			})
-			reset()
-		} else if (response.status == 400) {
-			const errors: { [field: string]: string } = await response.json()
-			for (const [field, message] of Object.entries(errors)) {
-				setError(field as keyof Station, { type: 'custom', message })
-			}
-		} else {
-			console.error('Unknown error:', await response.text())
-		}
-
-		return response
-	}
-
-  const afterAdd = (event: LeafletEvent) => setMarker(event.target)
-
-	return (
-    <Marker position={[station.latitude, station.longitude]} icon={pin} eventHandlers={{ add: afterAdd }}>
-			<Popup>
-				<form id="popup-add-form" className={`${styles.form}`} method="POST" onSubmit={handleSubmit(save)}>
-					<h3>Add New Marker</h3>
-					<label htmlFor="name">Name: </label>
-					<input type="text" required {...register('name')} />
-					<div className={styles.error}>{errors.name?.message}</div>
-					<div className={styles.location}>
-						<div>
-							<label htmlFor="latitude">Latitude: </label>
-							<input type="text" name="latitude" value={station.latitude.toFixed(5)} disabled />
-						</div>
-						<div>
-							<label htmlFor="longitude">Longitude: </label>
-							<input type="text" name="longitude" value={station.longitude.toFixed(5)} disabled />
-						</div>
-					</div>
-					<div className={styles['country-time']}>
-						<div className={styles.country}>
-							<label htmlFor="country">Country: </label>
-							<select className={emoji.className} {...register('country')}>
-								{countries.map((country) =>
-									<option className={emoji.className} key={country} value={country}>{flag(country)}</option>)}
-							</select>
-						</div>
-						<div className={styles.time}>
-							<label htmlFor="time_zone">Timezone: </label>
-							<input type="text" {...register('time_zone')} />
-						</div>
-					</div>
-					<label htmlFor="info_en">Additional information: </label>
-					<input type="text" {...register('info_en')} />
-					<button type="submit" disabled={isSubmitting}>
-						{isSubmitting ? <Loading text="Saving..." /> : 'Save'}
-					</button>
-				</form>
-			</Popup>
-		</Marker >
-	)
-}
-
-MapForm.displayName = 'MapForm'
 
 const GeoFootball = ({ stationsByCountry }: Properties) => {
 	const openStreetMaps = 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
@@ -170,7 +30,7 @@ const GeoFootball = ({ stationsByCountry }: Properties) => {
 
 	const [clusters, setClusters] = useState(stationsByCountry)
 
-  const remove = async (station: Station, index: number) => {
+	const remove = async (station: Station, index: number) => {
 		const response = await deleteStation(station.id)
 		if (response.ok) {
 			const cluster = clusters[station.country]
